Skip metrics that don't exist
If a metric doesn't exist, it should be skipped instead of returned. This is especially pertinant when running on certain storage nodes. For example, if running on the MGS node and the OSS metrics are not disabled, the collector would not export MGS metrics as it would return after the first OSS metric is not found.

Signed-Off-By: Robert Clark <robert.d.clark@hpe.com>