Create new metric type for BRW stats
A new metric description needed to be created for block read/write stats due to the various sub-labels for each metric. In the original 'constMetric' description, only the nodeName (OST, MGT or MDT) was a sub-label. For the various BRW stats to be included with this, a new metric would need to be created for every size and operation. For example, the 'discontiguous pages' metric with a size of '1' and operation of 'read' would require a different metric from 'discontiguous pages' with a size of '1' and operation of 'write'. The new 'brwMetric' description includes sub-labels for 'size' and 'operation' to allow similar metrics to be grouped together, minimizing clutter.

Signed-Off-By: Robert Clark <robert.d.clark@hpe.com>