Remove redundant metric prefixes
Originally, all metric names began with 'lustre_lustre_'. According to the documentation on BuildFQName, "BuildFQName joins the given three name components by '_'. Empty name components are ignored." Since all of the metrics in the lustre_exporter are in the 'lustre' namespace and a single subsystem, there is no reason to specify a subsystem. Fixes #11.

Signed-Off-By: Robert Clark <robert.d.clark@hpe.com>