Add 'health_check' metric
As written in the help text, nodes marked as 'healthy' will have a value
of 1 while nodes of any other status will be marked 0. Due to Prometheus
requiring a numeric value for all results, there isn't a readily available
way to parse unhealthy targets (such as an OST), hence the widespread
labelling of unhealthy.

Signed-Off-By: Robert Clark <robert.d.clark@hpe.com>