Consolidate CounterValue functions
Use of the variadic parameter allows us to specify an unknown number of labels
to the MustNewConstMetric function, which enables us to consolidate
CounterValue functions into a single unit. This will allow us to use only the
desired type without creating new functions for various types for metrics with
unique labels.

Signed-Off-By: Robert Clark <robert.d.clark@hpe.com>